refactor(board): share schedule helpers; flag overdue on the mobile rail
The kanban (#1242) colored its schedule chip by urgency, but the mobile task
rail still showed a plain always-amber chip, and both components duplicated
scheduleLabel/formatBoardDate.

- Extract scheduleLabel/formatBoardDate/scheduleUrgency into a shared
  src/lib/board-schedule.ts (single source of truth).
- Mobile rail now colors its schedule chip the same way: overdue (red, warning
  icon), due soon (amber), otherwise neutral — hydration-safe today, done cards
  never flag.
